More changes
Most are just typos and punctuation, but there's a section on spotters
where "Rock" was applied to several different races. I corrected it to
the appropriate race in each and just omitted it for the CIV one. I also
remove the "mighty" part from anything but the Rock because it only
seemed to fit them. The text for each could probably use a rewrite
customizing it for each race as it sounds a little awkward for anyone
but the Rock.